+++ lang/syn/src/codegen/accounts/__client_accounts.rs
@@ -193,7 +193,7 @@ pub fn generate(
             #(#re_exports)*
 
             #struct_doc
-            #[derive(anchor_lang::AnchorSerialize)]
+            #[derive(anchor_lang::AnchorSerialize, Debug, Default, Copy, Clone)]
             pub struct #name {
                 #(#account_struct_fields),*
             }